requirements.txt
anthropic>=0.34.0
beautifulsoup4>=4.12.3
elevenlabs>=1.3.0
fastapi>=0.115.0
openai>=1.55.3
python-dotenv>=1.0.1
redis>=5.0.8
requests>=2.32.3
pydantic>=2.9.2
uvicorn[standard]>=0.30.5

